Crates can be blown open safely

+++ code/game/objects/structures/crates_lockers/crates.dm
@@ -128,14 +128,27 @@
 /obj/structure/closet/crate/ex_act(severity)
 	switch(severity)
 		if(1.0)
-			for(var/obj/O in src.contents)
-				qdel(O)
+			// outpost 21 edit begin - stop destroying items
+			if(istype(src,/obj/structure/closet/crate/secure/loot))
+				for(var/obj/O in src.contents)
+					qdel(O)
+			else
+				for(var/obj/O in src.contents)
+					O.forceMove(get_turf(src))
+			// outpost 21 edit end
 			qdel(src)
 			return
 		if(2.0)
-			for(var/obj/O in src.contents)
-				if(prob(50))
+			// outpost 21 edit begin - stop destroying items
+			if(istype(src,/obj/structure/closet/crate/secure/loot))
+				for(var/obj/O in src.contents)
 					qdel(O)
+					if(prob(50))
+						qdel(O)
+			else
+				for(var/obj/O in src.contents)
+					O.forceMove(get_turf(src))
+			// outpost 21 edit end
 			qdel(src)
 			return
 		if(3.0)
